requirements.txt
appdirs==1.4.4
asgiref==3.5.2
certifi==2020.6.20
chardet==3.0.4
click==8.1.3
colorama==0.4.5
crispy-tailwind==0.5.0
distlib==0.3.5
Django==4.1
django-crispy-forms==1.14.0
django-taggit==3.0.0
django-widget-tweaks==1.4.8
filelock==3.7.1
gunicorn==20.1.0
idna==2.9
itsdangerous==2.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
platformdirs==2.5.2
protobuf==4.21.4
psycopg2==2.9.3
pycparser==2.20
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tailwind==3.1.5b0
tzdata==2022.1
urllib3==1.25.10
whitenoise==5.0.1
